+++ devicetypes/bangali/rooms-occupancy.src/rooms-occupancy.groovy
@@ -21,7 +21,7 @@
 *
 ***********************************************************************************************************************/
 
-public static String version()		{  return "v0.99.6"  }
+public static String version()		{  return "v1.0.0"  }
 private static boolean isDebug()	{  return false  }
 
 final String _SmartThings()	{ return 'ST' }
@@ -469,7 +469,6 @@ def on()	{
 		default:							break
 	}
 	switchOnOff(true)
-//	sendEvent(name: "switch", value: "on", descriptionText: "$device.displayName is on", displayed: true)
 }
 
 def setOnStateC(e)		{  state.onState = (e ? e.toString() : 'occupied')  }
@@ -477,7 +476,6 @@ def setOnStateC(e)		{  state.onState = (e ? e.toString() : 'occupied')  }
 def	off()		{
 	vacant()
 	switchOnOff(false)
-//	sendEvent(name: "switch", value: "off", descriptionText: "$device.displayName is off", displayed: true)
 }
 
 private switchOnOff(on)	{
@@ -507,435 +505,56 @@ def lock()		{  locked() }
 
 def unlock()	{  vacant()  }
 
-def occupied(hS = true, vM = false)			{  stateUpdateSetup('occupied', hS, vM); 		switchOnOff(true)  }
+def occupied(vM = false)		{  parent.occupied(vM)  }
 
-def checking(hS = true, vM = false)			{  stateUpdateSetup('checking', hS, vM)  }
+def checking(vM = false)		{  parent.checking(vM)  }
 
-def vacant(hS = true, vM = false)			{  stateUpdateSetup('vacant', hS, vM);			switchOnOff(false)  }
+def vacant(vM = false)			{  parent.vacant(vM)  }
 
-def donotdisturb(hS = true, vM = false)		{  stateUpdateSetup('donotdisturb', hS, vM);	switchOnOff(false)  }
+def donotdisturb(vM = false)	{  parent.donotdisturb(vM)  }
 
-def reserved(hS = true, vM = false)			{  stateUpdateSetup('reserved', hS, vM);		switchOnOff(false)  }
+def reserved(vM = false)		{  parent.reserved(vM)  }
 
-def asleep(hS = true, vM = false)			{  stateUpdateSetup('asleep', hS, vM);			switchOnOff(true)  }
+def asleep(vM = false)			{  parent.asleep(vM)  }
 
-def locked(hS = true, vM = false)			{  stateUpdateSetup('locked', hS, vM);			switchOnOff(false)  }
+def locked(vM = false)			{  parent.locked(vM)  }
 
-def engaged(hS = true, vM = false)			{  stateUpdateSetup('engaged', hS, vM);			switchOnOff(true)  }
+def engaged(vM = false)			{  parent.engaged(vM)  }
 
-def kaput(hS = true, vM = false)			{  stateUpdateSetup('kaput', hS, vM);			switchOnOff(false)  }
+def kaput(vM = false)			{  parent.kaput(vM)  }
 
-private stateUpdateSetup(rSt, hS, vM)	{
-	runIn(0, stateUpdate, [data: [newState:rSt, handleSwitches:hS, vacationMode:vM]])
-}
-
-def	stateUpdate(data)		{
-	if (!data)		return;
-	if (state.oldState != data.newState)		{
-        if (data.handleSwitches && parent)
-			setupTimer((int) (parent.handleSwitches(state.oldState, data.newState, true, data.vacationMode) ?: 0))
-		updateOccupancy(state.oldState, data.newState)
-		state.oldState = data.newState
-	}
-	resetTile(newState)
-}
-
-def updateOccupancy(oldOcc, newOcc) 	{
-	newOcc = newOcc?.toLowerCase()
-	def hT = getHubType()
-	def buttonMap = ['occupied':1, 'locked':4, 'vacant':3, 'reserved':5, 'checking':2, 'kaput':6, 'donotdisturb':7, 'asleep':8, 'engaged':9]
-	if (!newOcc || !(buttonMap.containsKey(newOcc)))	{
-		ifDebug("Missing or invalid parameter room occupancy: $newOcc")
-		return
-	}
-	sendEvent(name: "occupancy", value: newOcc, descriptionText: "$device.displayName changed to $newOcc", displayed: true)
-	if (hT == _Hubitat())		{
-		def img = "https://cdn.rawgit.com/adey/bangali/master/resources/icons/rooms${newOcc?.capitalize()}State.png"
-		sendEvent(name: "occupancyIconS", value: "<img src=$img height=25 width=25>", descriptionText: "$device.displayName $newOcc icon small", displayed: true)
-		sendEvent(name: "occupancyIconM", value: "<img src=$img height=50 width=50>", descriptionText: "$device.displayName $newOcc icon medium", displayed: true)
-		sendEvent(name: "occupancyIconL", value: "<img src=$img height=75 width=75>", descriptionText: "$device.displayName $newOcc icon large", displayed: true)
-		sendEvent(name: "occupancyIconXL", value: "<img src=$img height=100 width=100>", descriptionText: "$device.displayName $newOcc icon extra large", displayed: true)
-		sendEvent(name: "occupancyIconXXL", value: "<img src=$img height=150 width=150>", descriptionText: "$device.displayName $newOcc icon extra extra large", displayed: true)
-		sendEvent(name: "occupancyIconURL", value: img, descriptionText: "$device.displayName $newOcc icon URL", displayed: true)
-	}
-	def button = buttonMap[newOcc]
-	if (hT == _SmartThings())
-		sendEvent(name: "button", value: "pushed", data: [buttonNumber: button], descriptionText: "$device.displayName button $button was pushed.")
-	else
-		sendEvent(name:"pushed", value:button, descriptionText: "$device.displayName button $button was pushed.")
-
-	updateRoomStatusMsg()
-}
-
-private updateRoomStatusMsg()		{
-	def formatter = new java.text.SimpleDateFormat("EEE, MMM d yyyy @ h:mm:ss a z")
-	formatter.setTimeZone(location.timeZone)
-	state.statusMsg = formatter.format(now())
-	sendEvent(name: "status", value: state.statusMsg, displayed: false)
-}
-
+/*
 private	resetTile(occupancy)	{
 	sendEvent(name: occupancy, value: occupancy, descriptionText: "$device.displayName reset tile $occupancy", displayed: false)
 }
 
+*/
+
 def turnSwitchesAllOn()		{
-	if (parent)		{
-		parent.turnSwitchesAllOnOrOff(true)
-		if (getHubType() != _Hubitat())		updateSwitchInd(1);
-	}
+	if (parent)		parent.turnSwitchesAllOnOrOff(true);
+//		if (getHubType() != _Hubitat())		updateSwitchInd(1);
 }
 
 def turnSwitchesAllOff()		{
-	if (parent)		{
-		parent.turnSwitchesAllOnOrOff(false)
-		if (getHubType() != _Hubitat())		updateSwitchInd(0);
-	}
+	if (parent)		parent.turnSwitchesAllOnOrOff(false);
+//		if (getHubType() != _Hubitat())		updateSwitchInd(0);
 }
 
 def turnNightSwitchesAllOn()	{
  	ifDebug("turnNightSwitchesAllOn")
-	if (parent)	{
-		parent.dimNightLights()
-		if (getHubType() != _Hubitat())		updateNSwitchInd(1)
-	}
+	if (parent)		parent.dimNightLights();
+//		if (getHubType() != _Hubitat())		updateNSwitchInd(1)
 }
 
 def turnNightSwitchesAllOff()	{
 	ifDebug("turnNightSwitchesAllOff")
-	if (parent)		{
-		parent.nightSwitchesOff()
-		if (getHubType() != _Hubitat())		updateNSwitchInd(0)
-	}
+	if (parent)		parent.nightSwitchesOff();
+//		if (getHubType() != _Hubitat())		updateNSwitchInd(0)
 }
 
 def	turnOnAndOffSwitches()	{
 	if (parent)		parent.switchesOnOrOff();
-	setupTimer(-1)
-}
-
-def setupTimer(int timer)	{
-	if (timer != -1)	state.timerLeft = timer;
-	timerNext()
-}
-
-def timerNext()		{
-	int timerUpdate = (state.timerLeft > 30 ? 30 : (state.timerLeft < 5 ? state.timerLeft : 5))
-	def timerInd = (state.timerLeft > 3600 ? (state.timerLeft / 3600f).round(1) + 'h' : (state.timerLeft > 60 ? (state.timerLeft / 60f).round(1) + 'm' : state.timerLeft + 's')).replace(".0","")
-	if (getHubType() != _Hubitat())
-		sendEvent(name: "timer", value: (timerInd ?: '--'), displayed: false)
-	else
-		sendEvent(name: "countdown", value: timerInd, descriptionText: "countdown timer: $timerInd", displayed: true)
-	state.timerLeft = state.timerLeft - timerUpdate
-	(state.timerLeft > 0 ? runIn(timerUpdate, timerNext) : unschedule('timerNext'))
+//	setupTimer(-1)
 }
 
 private ifDebug(msg = null, level = null)	{  if (msg && (isDebug() || level == 'error'))	log."${level ?: 'debug'}" " $device.displayName device: " + msg  }
-
-//
-// REMOVE THE FOLLOWING CODE FOR HUBITAT TILL END OF FILE		<<<<<
-//
-
-def updateMotionInd(motionOn)		{
-	def vV = 'none'
-	def dD = "indicate no motion sensor"
-	switch(motionOn)	{
-		case 1:		vV = 'active';		dD = "indicate motion active";		break
-		case 0:		vV = 'inactive';	dD = "indicate motion inactive";	break
-	}
-	sendEvent(name: 'motionInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateLuxInd(lux)		{
-	sendEvent(name: 'luxInd', value: (lux == -1 ? '--' : (lux <= 100 ? lux : formatNumber(lux))), descriptionText: (lux == -1 ? "indicate no lux sensor" : "indicate lux value"), displayed: false)
-}
-
-def updateHumidityInd(humidity)		{
-	sendEvent(name: 'humidityInd', value: (humidity == -1 ? '--' : humidity.toString() + '%'), descriptionText: (humidity == -1 ? "indicate no humidity sensor" : "indicate humidity value"), displayed: false)
-}
-def updateContactInd(contactClosed)		{
-	def vV = 'none'
-	def dD = "indicate no contact sensor"
-	switch(contactClosed)	{
-		case 1:		vV = 'closed';	dD = "indicate contact closed";		break
-		case 0:		vV = 'open';	dD = "indicate contact open";		break
-	}
-	sendEvent(name: 'contactInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateContactRTIndC(contactStatus)		{
-	def vV = 'none'
-	def dD = "indicate no contact sensor"
-	switch(contactStatus)	{
-		case 1:		vV = 'closed';	dD = "indicate contact closed";		break
-		case 0:		vV = 'open';	dD = "indicate contact open";		break
-	}
-	sendEvent(name: 'contactRTInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateSwitchInd(switchOn)		{
-	def vV = '--'
-	def dD = "indicate no switches to turn on in room"
-	switch(switchOn)	{
-		case 1:		vV = 'on';	dD = "indicate at least one switch in room is on";	break
-		case 0:		vV = 'off';	dD = "indicate all switches in room is off";		break
-	}
-	sendEvent(name: 'switchInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updatePresenceInd(presencePresent)		{
-	def vV = 'none'
-	def dD = "indicate no presence sensor"
-	switch(presencePresent)		{
-		case 1:		vV = 'present';	dD = "indicate presence present";		break
-		case 0:		vV = 'absent';	dD = "indicate presence not present";	break
-	}
-	sendEvent(name: 'presenceInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updatePresenceActionInd(presenceAction)		{
-	def vV = '--'
-	def dD = "indicate no presence sensor"
-	switch(presenceAction)	{
-		case '1':		vV = 'Arrival';		dD = "indicate arrival action when present";						break
-		case '2':		vV = 'Departure';	dD = "indicate departure action when not present";					break
-		case '3':		vV = 'Both';		dD = "indicate both arrival and depature action with presence";		break
-		case '4':		vV = 'Neither';		dD = "indicate no action with with present";						break
-	}
-	sendEvent(name: 'presenceActionInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updatePresenceEngagedInd(presenceEngaged)		{
-	sendEvent(name: 'presenceEngagedInd', value: (presenceEngaged == -1 ? '--' : presenceEngaged), descriptionText: (presenceEngaged == -1 ? "indicate no presence sensor" : "indicate if presence action continuous"), displayed: false)
-}
-
-def updateBusyEngagedInd(busyEngaged)		{
-	sendEvent(name: 'busyEngagedInd', value: (busyEngaged == -1 ? '--' : "$busyEngaged\ntraffic"), descriptionText: (busyEngaged == -1 ? "indicate no presence sensor" : "indicate traffic check"), displayed: false)
-}
-
-def updateDoWInd(dow)		{
-	def		vV
-	switch(dow)	{
-		case '1':	vV = 'Monday';		break
-		case '2':	vV = 'Tuesday';		break
-		case '3':	vV = 'Wednesday';	break
-		case '4':	vV = 'Thursday';	break
-		case '5':	vV = 'Friday';		break
-		case '6':	vV = 'Saturday';	break
-		case '7':	vV = 'Sunday';		break
-		case '8':	vV = 'M - F';		break
-		case '9':	vV = 'S & S';		break
-		default:	vV = 'Everyday';	break
-	}
-	sendEvent(name: 'dowInd', value: vV, descriptionText: "indicate run on only these days of the week: $vV", displayed: false)
-}
-
-def updateTimeInd(timeFromTo)		{
-	sendEvent(name: 'timeInd', value: timeFromTo, descriptionText: "indicate time from to", displayed: false)
-}
-
-def updateTemperatureInd(temp)		{
-	def tS = (location.temperatureScale ?: 'F')
-	sendEvent(name: 'temperatureInd', value: (temp == -1 ? '--' : temp + '°' + tS), unit: tS, descriptionText: (temp == -1 ? "indicate no temperature sensor" : "indicate temperature value"), displayed: false)
-}
-
-def updateOutTempIndC(temp)		{
-	def tS = (location.temperatureScale ?: 'F')
-	sendEvent(name: 'outTempInd', value: (temp == -1 ? '--' : temp + '°' + tS), unit: tS, descriptionText: (temp == -1 ? "indicate no temperature sensor" : "indicate temperature value"), displayed: false)
-}
-
-def updateVentIndC(vent)		{
-	def vL, dD
-	switch(vent)	{
-		case -1:	vL = 'none';		dD = "indicate no vents";		break
-		case 0:		vL = 'closed';		dD = "indicate vents closed";	break
-		default:	vL = 'open';		dD = "indicate vent open";		break
-	}
-//	sendEvent(name: 'ventInd', value: (vL == 'open' ? formatNumber(vent) : vL), descriptionText: dD, displayed: false)
-	sendEvent(name: 'ventInd', value: vL, descriptionText: dD, displayed: false)
-}
-
-def updateMaintainIndC(maintainRoomTemp, turnOn, temperature, coolTemp, heatTemp, outTempSensor, outTemp, autoAdjustWithOutdoor)		{
-	def temp = 999
-	if (turnOn)		{
-		boolean isFarenheit = (location.temperatureScale == 'F' ? true : false)
-		if (outTempSensor && autoAdjustWithOutdoor)	{
-			if (outTemp > (isFarenheit ? 90 : 26.7))	coolTemp = coolTemp - (isFarenheit ? 0.5 : 0.28);
-			if (outTemp < (isFarenheit ? 32 : 0))		heatTemp = heatTemp + (isFarenheit ? 0.5 : 0.28);
-		}
-		if (maintainRoomTemp == '1')
-			temp = coolTemp
-		else if (maintainRoomTemp == '2')
-			temp = heatTemp
-		else if (maintainRoomTemp == '3' || maintainRoomTemp == '5')	{
-			def x = Math.abs(temperature - coolTemp)
-			def y = Math.abs(temperature - heatTemp)
-			temp = (x >= y ? heatTemp : coolTemp)
-		}
-	}
-	else
-		temp = -1
-	if (temp != 999)	{
-		def tS = (location.temperatureScale ?: 'F')
-		sendEvent(name: 'maintainInd', value: (temp == -1 ? '--' : temp + '°' + tS), unit: tS, descriptionText: (temp == -1 ? "indicate no maintain temperature" : "indicate maintain temperature value"), displayed: false)
-	}
-}
-
-def updateThermostatIndC(currentPresence, maintainRoomTemp, useThermostat, currentThermostatOperatingState, roomCoolSwitch, roomHeatSwitch)		{
-	def thermo = 9
-	def isHere = currentPresence.contains('present')
-	if ((useThermostat && currentThermostatOperatingState == 'cooling') || (!useThermostat && roomCoolSwitch == on))
-		thermo = 4
-	else if ((useThermostat && currentThermostatOperatingState == 'heating') || (!useThermostat && roomHeatSwitch == on))
-		thermo = 5
-	else if (!isHere && ['1', '2', '3'].contains(maintainRoomTemp))
-		thermo = 0
-	else if (maintainRoomTemp == '3')
-		thermo = 1
-	else if (maintainRoomTemp == '1')
-		thermo = 2
-	else if (maintainRoomTemp == '2')
-		thermo = 3
-	ifDebug("updateThermostatIndC: thermo: $thermo")
-	def vV = '--'
-	def dD = "indicate no thermostat setting"
-	switch(thermo)	{
-		case 0:		vV = 'off';			dD = "indicate thermostat not auto";	break
-		case 1:		vV = 'auto';		dD = "indicate thermostat auto";		break
-		case 2:		vV = 'autoCool';	dD = "indicate thermostat auto cool";	break
-		case 3:		vV = 'autoHeat';	dD = "indicate thermostat auto heat";	break
-		case 4:		vV = 'cooling';		dD = "indicate thermostat cooling";		break
-		case 5:		vV = 'heating';		dD = "indicate thermostat heating";		break
-	}
-	sendEvent(name: 'thermostatInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateThermoOverrideIndC(thermoOverride)		{
-	sendEvent(name: 'thermoOverrideInd', value: thermoOverride, descriptionText: "indicate thermo override minutes", displayed: false)
-}
-
-def updateFanIndC(fan)		{
-	def vV = '--'
-	def dD = "indicate no fan setting"
-	switch(fan)	{
-		case 0:		vV = 'off';		dD = "indicate fan off";				break
-		case 1:		vV = 'low';		dD = "indicate fan on at low speed";	break
-		case 2:		vV = 'medium';	dD = "indicate fan on at medium speed";	break
-		case 3:		vV = 'high';	dD = "indicate fan on at high speed";	break
-	}
-	sendEvent(name: 'fanInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateRulesInd(rules)		{
-	sendEvent(name: 'rulesInd', value: (rules == -1 ? '0' : rules), descriptionText: (rules == -1 ? "indicate no rules" : "indicate rules count"), displayed: false)
-}
-
-def updateLastRuleInd(rule)		{
-	sendEvent(name: 'lastRuleInd', value: (rule == -1 ? '--' : rule), descriptionText: (rule == -1 ? "indicate no rule executed" : "indicate rule number last executed"), displayed: false)
-}
-
-def updatePauseInd(pMode)		{
-	sendEvent(name: 'pauseInd', value: (pMode == -1 ? '--' : pMode), descriptionText: (pMode == -1 ? "indicate no pause modes" : "indicate pause modes"), displayed: false)
-}
-
-def updatePowerInd(power)		{
-	sendEvent(name: 'powerInd', value: (power == -1 ? '--' : (power <= 100 ? power : formatNumber(power))), descriptionText: (power == -1 ? "indicate no power sensor" : "indicate power value"), displayed: false)
-}
-
-def updateEWattsInd(eWatts)		{
-	sendEvent(name: 'eWattsInd', value: (eWatts == -1 ? '--' : (eWatts <= 100 ? eWatts : formatNumber(eWatts))), descriptionText: (eWatts == -1 ? "indicate no engaged watts" : "indicate engaged watts value"), displayed: false)
-}
-
-def updateAWattsInd(aWatts)		{
-	sendEvent(name: 'aWattsInd', value: (aWatts == -1 ? '--' : (aWatts <= 100 ? aWatts : formatNumber(aWatts))),
-				descriptionText: (aWatts == -1 ? "indicate no asleep watts" : "indicate asleep watts value"), displayed: false)
-}
-
-def updateESwitchInd(switchOn)		{
-	def vV = '--'
-	def dD = "indicate no engaged switch"
-	switch(switchOn)	{
-		case 1:		vV = 'on';	dD = "indicate engaged switch is on";	break
-		case 0:		vV = 'off';	dD = "indicate engaged switch is off";	break
-	}
-	sendEvent(name: 'eSwitchInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateTimersInd(noMotion, dimTimer, noMotionEngaged, noMotionAsleep)		{
-	sendEvent(name: 'noMotionInd', value: (noMotion ? formatNumber(noMotion) : '--'), descriptionText: (noMotion ? "indicate motion timer for occupied state" : "indicate no motion timer for occupied state"), displayed: false)
-	sendEvent(name: 'dimTimerInd', value: (dimTimer ? formatNumber(dimTimer) : '--'), descriptionText: (dimTimer ? "indicate timer for checking state" : "indicate no timer for checking state"), displayed: false)
-	sendEvent(name: 'noMotionEngagedInd', value: (noMotionEngaged ? formatNumber(noMotionEngaged) : '--'), descriptionText: (noMotionEngaged ? "indicate motion timer for engaged state" : "indicate no motion timer for engaged state"), displayed: false)
-	sendEvent(name: 'noMotionAsleepInd', value: (noMotionAsleep ? formatNumber(noMotionAsleep) : '--'), descriptionText: (noMotionAsleep ? "indicate motion timer for asleep state" : "indicate no motion timer for asleep state"), displayed: false)
-}
-
-def updateOSwitchInd(switchOn)		{
-	def vV = '--'
-	def dD = "indicate no occupied switches"
-	switch(switchOn)	{
-		case 1:		vV = 'on';	dD = "indicate at least one occupied switch is on";		break
-		case 0:		vV = 'off';	dD = "indicate all occupied switches is off";			break
-	}
-	sendEvent(name: 'oSwitchInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateASwitchInd(switchOn)		{
-	def vV = '--'
-	def dD = "indicate no asleep switches"
-	switch(switchOn)	{
-		case 1:		vV = 'on';	dD = "indicate at least one asleep switch is on";	break
-		case 0:		vV = 'off';	dD = "indicate all asleep switches are off";		break
-	}
-	sendEvent(name: 'aSwitchInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateNSwitchInd(switchOn)		{
-	def vV = '--'
-	def dD = "indicate no night switches"
-	switch(switchOn)	{
-		case 1:		vV = 'on';	dD = "indicate at least one night switch is on";	break
-		case 0:		vV = 'off';	dD = "indicate all night switches are off";			break
-	}
-	sendEvent(name: 'nSwitchInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateLSwitchInd(switchOn)		{
-	def vV = '--'
-	def dD = "indicate no locked switch"
-	switch(switchOn)	{
-		case 1:		vV = 'on';	dD = "indicate locked switch is on";	break
-		case 0:		vV = 'off';	dD = "indicate locked switch is off";	break
-	}
-	sendEvent(name: 'lSwitchInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-def updateTurnAllOffInd(turnOff)		{
-	sendEvent(name: 'turnAllOffInd', value: turnOff, descriptionText: "indicate if all switches should be turned off when no rules match", displayed: false)
-}
-
-def updateDimByLevelInd(dimBy, dimTo)		{
-	sendEvent(name: 'dimByLevelInd', value: (dimBy == -1 && dimTo == -1 ? '-- / --' : "${(dimBy == -1 ? '--' : dimBy + '%')} /\n${(dimTo == -1 ? '--' : dimTo + '% ')}"), descriptionText: (dimBy == -1 && dimTo == -1 ? "indicate no dimming" : "indicate dimming by / to level"), displayed: false)
-}
-
-def updateAdjRoomsInd(aRooms)		{
-	sendEvent(name: 'aRoomInd', value: (aRooms == -1 ? '--' : aRooms + '\nrooms'), descriptionText: (aRooms == -1 ? "indicate no adjacent rooms" : "indicate how many adjacent rooms"), displayed: false)
-}
-
-def updateWSSInd(wSS)		{
-	sendEvent(name: 'wSSInd', value: (wSS == -1 ? '--' : wSS), descriptionText: (wSS == -1 ? "indicate no window shades" : "indicate window shade position"), displayed: false)
-}
-
-def updateAdjMotionInd(motionOn)		{
-	def vV = 'none'
-	def dD = "indicate no adjacent motion sensor"
-	switch(motionOn)	{
-		case 1:		vV = 'active';		dD = "indicate adjacent motion active";		break
-		case 0:		vV = 'inactive';	dD = "indicate adjacent motion inactive";	break
-	}
-	sendEvent(name: 'aMotionInd', value: vV, descriptionText: dD, displayed: false)
-}
-
-private formatNumber(number)	{
-	int n = number as Integer
-	return (n > 0 ? String.format("%,d", n) : '')
-}
-
-// REMOVE TILL HERE FOR HUBITAT		<<<<<